+++ src/main/java/com/cimadev/cimpleWaypointSystem/command/tpa/TpaCommand.java
@@ -10,9 +10,6 @@ import net.minecraft.command.argument.EntityArgumentType;
 import net.minecraft.server.command.CommandManager;
 import net.minecraft.server.command.ServerCommandSource;
 import net.minecraft.server.network.ServerPlayerEntity;
-import net.minecraft.text.ClickEvent;
-import net.minecraft.text.HoverEvent;
-import net.minecraft.text.Style;
 import net.minecraft.text.Text;
 import net.minecraft.util.Formatting;
 
@@ -47,62 +44,9 @@ public class TpaCommand {
             );
             return 0;
         }
-        TeleportRequestManager.getInstance().addRequest(new TeleportRequest(
-                origin,
-                target,
-                false
-        ));
-        target.sendMessage(
-                Text.literal("")
-                .append(origin.getName().copy().formatted(Colors.PLAYER))
-                .append(" has requested to teleport to you. Type ")
-                .append(
-                        Text.literal("/tpaccept")
-                                .setStyle(
-                                        Style.EMPTY
-                                        .withClickEvent(new ClickEvent(
-                                                ClickEvent.Action.SUGGEST_COMMAND,
-                                                "/tpaccept"
-                                        ))
-                                        .withHoverEvent(new HoverEvent(
-                                            HoverEvent.Action.SHOW_TEXT,
-                                            Text.literal("Click here to accept")
-                                        ))
-                                )
-                                .formatted(Formatting.YELLOW)
-                )
-                .append(" to accept the request or ")
-                .append(
-                        Text.literal("/tpdeny")
-                                .setStyle(
-                                        Style.EMPTY
-                                        .withClickEvent(new ClickEvent(
-                                                ClickEvent.Action.SUGGEST_COMMAND,
-                                                "/tpdeny"
-                                        ))
-                                        .withHoverEvent(new HoverEvent(
-                                                HoverEvent.Action.SHOW_TEXT,
-                                                Text.literal("Click here to deny")
-                                        ))
-                                )
-                                .formatted(Formatting.RED)
-                )
-                .append(" to deny it. The request will expire in ")
-                .append(
-                        Text.literal(
-                                Long.toString(TeleportRequestManager.getInstance().getRequestTTL()/20)
-                        )
-                                .append(" seconds")
-                                .formatted(Colors.TIME)
-                )
-                .formatted(Colors.DEFAULT)
-        );
-        origin.sendMessage(
-                Text.literal("")
-                .append(target.getName().copy().formatted(Colors.PLAYER))
-                .append(" has received your teleport request.")
-                .formatted(Colors.DEFAULT)
-        );
+        final TeleportRequest request = new TeleportRequest(origin, target, false);
+        TeleportRequestManager.getInstance().addRequest(request);
+        TpaMessages.sendRequestMessages(request);
 
         return 1;
     }
